feat: Focus headings after changing page
When a user using a screen reader navigates the webshop, the heading
should be focused when changing page. This is more inline with the
guidelines for WGAT 2.0 2.4.3.

This was made possible by using the domFocus helper from Elm. The
focusable elements needs to have an id, and a tab-index too be
focusable. The tab-index is set to -1 to not be accessible during normal
tabbing.

Changed message-box aria-live to only be assertive when the status level
is Error or Warning. Also removed aria-live from buttons, which are no
longer assertive if disabled or loading.