feat: Improve radio buttons accessibility
Made the hidden original radio button input not selectable for screen
reader cursor, making the label not be read twice.

Also removed required Bool parameter for radio onCheck event, as it was
uneccesary since a radio button may never be unchecked.

Added an optional aria label which may be passed to a radio button.