feat: Remove empty notifications div
When there was no notifications, the parent div was still present with
its paddings. This parent div blocked half the header, making the header
logo and hamburger button hard to click on mobile. This was especially
present on older Android versions, where the hamburger menu was almost
impossible to click.