feat: Screen reader improvements for Group expandables (#387)
The group section was a h3 header with a button as the child element.
This was not handled well by all screen readers, where some were
confused if the focused item was a header or a button. Changed the group
sections to only be buttons, and not headers.

Also some other small improvements regarding hidden fieldset legends not
being focusable by screen reader, and including the radio button
subtitle in its aria label.