fix: Time zone error on fare contract
Removed usage of the FareTime object which where parsed in js. The
FareTime object did not take into consideration the current timezone. It
could have been modified to work with timezone, but I found it easier to
just use epoch millis on the model objects, and format them correctly in
the views where they are used.