fix: Time zone error on fare contract (#435)
Removed usage of the FareTime object which where parsed in js. The
FareTime object did not take into consideration the current timezone. It
could have been modified to work with timezone, but I found it easier to
just use epoch millis on the model objects, and format them correctly in
the views where they are used.

Note that there is still a bug as all dates are parsed with the browser's
current timezone. If parsing a summer time date during winter time, the
parsed date will be one hour off the actual date when parsed in summer
time.